Closing issue #29 - Graphics tearing.
In commit 48323fi, when method quickDivide() has been removed, the two
values divided that were passed as arguments to the method were kept
as-is, having a coma (,) instead of a divide sign (/), thus, making the
result not a division.

Changed that coma to a divide sign and graphic problem (observed in Out
of this World) are now solved.