Add memory capacity option
This allows the system to have upto 16MB of combined DRAM and
VRAM. This is really only useful for homebrew and it appears that not
all retail software will work properly if configured with non-stock
values.

This change also includes a reworking of the libretro variable
handling, tweaks to some options and descriptions, as well as some
changes to MADAM CEL rendering to accomidate for variable memory.